segment-anything-2-real-time-main/requirements.txt
tqdm
einops
moviepy
timm
transformers
supervision
opencv-python
pytest
natsort
hydra-core
iopath
pillow 